fix: better removal of alternate versions
Improved the logic where we split the alternate versions in the merge versions manager to also account for local alternate versions. Since we're only operating on Shoko managed libraries then I'm fine with the plugin completely managing the alternate versions so we'll just take control of it all, instead of the partway control we before this commit where we were intentionally ignoring the local versions and somewhat half-baked removing the linked local versions.

Also hello 👋 to the one reading this.