chore: add concurrency lock to API client
Added a concurrency lock to the API client, to allow a maximum of 10 concurrent requests at all times, as to not overload the server with a DoS attack when iterating a long-running series like e.g. One Piece.

Also tweaked the log statements and made POST requests able to be cached if needed, but they skip the cache by default.