feat: add iterative file checks mode in VFS
Added a new iterative file check mode for VFS generation, to be used when iterating the file system is expensive (e.g. if the libraries are on a network mapped mount/drive, etc.), if you just prefer it to be done incrementally instead of up front. The setting is marked as debug.