misc: allow syncing vfs thread count to refresh thread count
Allow synchronizing the number of threads to use during a VFS (re-)generation to the number of threads to use during a recursive refresh of a folder (any folder, by it the library itself, a series within, etc.) defined in the jellyfin core system settings by setting the VFS threads setting to `-1`. Setting it to `0` or anything below `-1` will still set it to use the number of available CPU cores, but `-1` is now reserved for following the system setting.