refactor: always send language code and actual rating to API clients
Always send the proper language code and actual rating to the API clients. We determine if it's likely to be an API client by tracking the currently in-flight requests to the remote images endpoint (like we've always done so far), but now e're also taking note of the item's id to check if we get a call to the image provider with said id. It's not fool proof, but it should work in 99.9% of use-cases unless you happen to be fetching the images for an item that's also currently getting an automated image refresh, in which case, good luck timing that.